pyproject.toml
[tool.poetry]
name = "picture_lib_utils"
version = "0.0.1"
description = ""
authors = ["James Boland <james@jboland.tech>"]

[tool.poetry.dependencies]
python = "^3.10"
Pillow = "^9.2.0"
click = "^8.1.3"

[tool.poetry.dev-dependencies]

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
